projeto-individual-2/vinicius-muniz/projeto-2/skin-cancer-mlflow/requirements.txt
mlflow>=2.17,<3.0
transformers>=4.44
torch>=2.5
torchvision>=0.20
Pillow>=10
pandas>=2.2
numpy>=1.26,<3
scikit-learn>=1.5
fastapi>=0.115
uvicorn>=0.30
python-multipart>=0.0.12
pyyaml>=6
requests>=2.32
